requirements.txt
acryl-datahub==0.13.1.2
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.3.0
apicurioregistrysdk==2.5.10
async-timeout==4.0.3
attrs==23.2.0
avro==1.11.3
avro-gen3==0.7.12
cached-property==1.5.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-default-group==1.2.4
click-spinner==0.1.10
confluent-kafka==2.3.0
Deprecated==1.2.14
docker==7.0.0
exceptiongroup==1.2.0
expandvars==0.12.0
fastavro==1.9.4
frozenlist==1.4.1
grpcio==1.62.1
grpcio-tools==1.62.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.4
httpx==0.27.0
humanfriendly==10.0
hyperframe==6.0.1
idna==3.6
ijson==3.2.3
importlib_metadata==7.1.0
iniconfig==2.0.0
jsonref==1.1.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
microsoft-kiota-abstractions==0.6.0
microsoft-kiota-http==0.4.4
microsoft-kiota-serialization-json==0.3.7
microsoft-kiota-serialization-text==0.2.1
mixpanel==4.10.1
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
packaging==24.0
pluggy==1.4.0
progressbar2==4.4.2
protobuf==4.25.3
psutil==5.9.8
pydantic==1.10.14
pytest==8.1.1
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
python-utils==3.8.2
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-file==2.0.0
rpds-py==0.18.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
sentry-sdk==1.43.0
six==1.16.0
sniffio==1.3.1
tabulate==0.9.0
termcolor==2.4.0
testcontainers==4.1.1
toml==0.10.2
tomli==2.0.1
typing-inspect==0.9.0
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.1
